+++ plugins/source/aws/client/client.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"reflect"
 	"strings"
 	"sync"
 
@@ -21,7 +22,7 @@ import (
 type Client struct {
 	// Those are already normalized values after configure and this is why we don't want to hold
 	// config directly.
-	ServicesManager ServicesManager
+	ServicesManager *ServicesManager
 	logger          zerolog.Logger
 	// this is set by table clientList
 	AccountID            string
@@ -91,7 +92,7 @@ func (s *ServicesManager) InitServicesForPartitionAccount(partition, accountId s
 
 func NewAwsClient(logger zerolog.Logger, spec *Spec) Client {
 	return Client{
-		ServicesManager: ServicesManager{
+		ServicesManager: &ServicesManager{
 			services: ServicesPartitionAccountMap{},
 		},
 		logger:       logger,
@@ -112,22 +113,35 @@ func (c *Client) ID() string {
 		string(c.WAFScope),
 		c.LanguageCode,
 	}
-
 	return strings.TrimRight(strings.Join(idStrings, ":"), ":")
 }
 
 func (c *Client) updateService(service AWSServiceName) {
+	// Check to see if the service is already initialized
+	svc := c.ServicesManager.ServicesByPartitionAccount(c.Partition, c.AccountID).GetService(service)
+	if svc != nil {
+		return
+	}
+	// If service is not  initialized, lock the account mutex and check again
 	c.accountMutex[c.AccountID].Lock()
 	defer c.accountMutex[c.AccountID].Unlock()
-	c.ServicesManager.ServicesByPartitionAccount(c.Partition, c.AccountID).InitService(c.AWSConfig, service)
+	svc = c.ServicesManager.ServicesByPartitionAccount(c.Partition, c.AccountID).GetService(service)
+	// if service is still not initialized, initialize it
+	if svc == nil {
+		c.logger.Debug().Msgf("updating service %s for: %s", service.String(), c.AccountID)
+		c.ServicesManager.ServicesByPartitionAccount(c.Partition, c.AccountID).InitService(c.AWSConfig, service)
+	}
 }
 func (c *Client) Services(service_names ...AWSServiceName) *Services {
 	for _, service := range service_names {
-		if c.ServicesManager.ServicesByPartitionAccount(c.Partition, c.AccountID).GetService(service) == nil {
-			c.updateService(service)
-		}
+		c.updateService(service)
 	}
 	return c.ServicesManager.ServicesByPartitionAccount(c.Partition, c.AccountID)
+}
+
+func (s *Services) Duplicate() Services {
+	duplicateServices := *s
+	return duplicateServices
 }
 
 func (c *Client) Duplicate() *Client {
@@ -196,6 +210,18 @@ func Configure(ctx context.Context, logger zerolog.Logger, spec Spec) (schema.Cl
 		return nil, fmt.Errorf("spec validation failed: %w", err)
 	}
 	spec.SetDefaults()
+
+	if spec.TableOptions != nil {
+		structVal := reflect.ValueOf(*spec.TableOptions)
+		fieldNum := structVal.NumField()
+		for i := 0; i < fieldNum; i++ {
+			field := structVal.Field(i)
+			if field.IsValid() && !field.IsZero() {
+				logger.Warn().Msg("table_options is deprecated and will be removed soon. Please reach out to the CloudQuery team if you require this feature")
+				break
+			}
+		}
+	}
 
 	client := NewAwsClient(logger, &spec)