+++ plugins/source/github/client/client.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/beatlabs/github-auth/app/inst"
@@ -14,6 +15,7 @@ import (
 	"github.com/google/go-github/v49/github"
 	"github.com/rs/zerolog"
 	"golang.org/x/oauth2"
+	"golang.org/x/sync/errgroup"
 )
 
 type Client struct {
@@ -116,7 +118,7 @@ func New(ctx context.Context, logger zerolog.Logger, spec Spec) (schema.ClientMe
 		repos:       spec.Repos,
 	}
 	c.logger.Info().Msg("Discovering repositories")
-	orgRepositories, err := c.discoverRepositories(ctx, spec.Orgs, spec.Repos)
+	orgRepositories, err := c.discoverRepositories(ctx, spec.DiscoveryConcurrency, spec.Orgs, spec.Repos)
 	if err != nil {
 		return nil, fmt.Errorf("failed to discover repositories: %w", err)
 	}
@@ -136,25 +138,43 @@ func servicesForClient(c *github.Client) GithubServices {
 	}
 }
 
-func (c *Client) discoverRepositories(ctx context.Context, orgs []string, repos []string) (map[string][]*github.Repository, error) {
-	opts := &github.RepositoryListByOrgOptions{ListOptions: github.ListOptions{PerPage: 100}}
-
+func (c *Client) discoverRepositories(ctx context.Context, discoveryConcurrency int, orgs []string, repos []string) (map[string][]*github.Repository, error) {
 	orgRepos := make(map[string][]*github.Repository)
-	for _, org := range orgs {
-		services := c.servicesForOrg(org)
-		for {
-			repos, resp, err := services.Repositories.ListByOrg(ctx, org, opts)
-			if err != nil {
-				return nil, err
-			}
-			orgRepos[org] = append(orgRepos[org], repos...)
+	orgReposLock := sync.Mutex{}
+	errorGroup, gtx := errgroup.WithContext(ctx)
+	errorGroup.SetLimit(discoveryConcurrency)
 
-			if resp.NextPage == 0 {
-				break
+	for _, org := range orgs {
+		org := org
+		opts := &github.RepositoryListByOrgOptions{ListOptions: github.ListOptions{PerPage: 100}}
+		services := c.servicesForOrg(org)
+		errorGroup.Go(func() error {
+			orgRepositories := []*github.Repository{}
+			for {
+				repos, resp, err := services.Repositories.ListByOrg(gtx, org, opts)
+				if err != nil {
+					return err
+				}
+				orgRepositories = append(orgRepositories, repos...)
+
+				if resp.NextPage == 0 {
+					break
+				}
+				opts.Page = resp.NextPage
 			}
-			opts.Page = resp.NextPage
-		}
+
+			orgReposLock.Lock()
+			defer orgReposLock.Unlock()
+			orgRepos[org] = orgRepositories
+
+			return nil
+		})
 	}
+
+	if err := errorGroup.Wait(); err != nil {
+		return nil, err
+	}
+
 	seenOrgs := make(map[string]struct{})
 	for _, repo := range repos {
 		repoSplit := splitRepo(repo)